+++ lib/testing-utils/src/factories/index.ts
@@ -1,5 +1,6 @@
 import assessment from "./assessment";
 import dimension from "./dimension";
 import hash from "./hash";
+import range from "./range";
 
-export default { assessment, dimension, hash }
+export default { assessment, dimension, hash, range }
