--- lib/testing-utils/src/factory/assessment.ts
@@ -1,3 +1,0 @@
-const assessment = {}
-
-export default assessment;
